move action table components to render
